backend/requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.4.1
email-validator==2.2.0
marshmallow==3.23.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
simple-websocket==1.1.0
